Remove CLI documentation generation scripts
This commit deletes the following files related to the generation of CLI documentation:
- hack/generate-cli-docs/generate_markdown.go
- hack/generate-cli-docs/generate_root_index.go
- hack/generate-cli-docs/main.go
- hack/generate-cli-docs/map.go
- hack/generate-cli-docs/util.go

These files were responsible for creating markdown documentation for the CLI commands and their options. The removal of these files indicates a shift in the documentation strategy or a refactor of the documentation generation process.

On-behalf-of: Gerald Morrison (SAP) <gerald.morrison@sap.com>
Signed-off-by: Gerald Morrison (SAP) <gerald.morrison@sap.com>