feat: implement documentation versioning and update header for versio… (#487)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
The PR introduces a version toggle on the website. It is using a
versions.json in the data/ folder that contains all available versions.
New versions of the website need to be cut off from the main branch and
afterwards built using the baseURL + the new version and merge this into
the public/ folder. This holds all versions in parallel, accessible by
introducing a v2 in the URL, since Gh pages are not able to use multiple
branches.

```
hugo build  --baseURL https://ocm.software/v2/ \
     --destination ./public/v2 \
     --minify --gc
```

The PR will introduce the version toggle and the version "current" which
will live on the root level in the public/ folder that will be
published.

Related to
https://github.com/open-component-model/ocm-project/issues/482

Signed-off-by: Gerald Morrison (SAP) <gerald.morrison@sap.com>
Co-authored-by: Gerald Morrison (SAP) <gerald.morrison@sap.com>