chore: Fix the html checker (#488)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
The current htmlcheck is not able to successfully scan links on the real
website, so it produces quite a bunch of false positives. This PR adds
the link check using
[`linkinator`](https://github.com/JustinBeckwith/linkinator). Both
together should now really guarantee that there are no broken links as
we always scan the real website generated as deploy preview by Netlify.

---------

Signed-off-by: Gerald Morrison (SAP) <gerald.morrison@sap.com>
Co-authored-by: Gerald Morrison (SAP) <gerald.morrison@sap.com>