chore: move out go dependencies to hack folder and update module versions (#526)
On-behalf-of: Gerald Morrison (SAP) <gerald.morrison@sap.com>

<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it
The website should not have direct Go dependencies other than required
for Hugo modules. Therefore we move out the dependencies into an own
module required for the small program in the hack folder.

The thulite package and the Hugo server package was updated to a later
version.

---------

Signed-off-by: Gerald Morrison (SAP) <gerald.morrison@sap.com>
Co-authored-by: Gerald Morrison (SAP) <gerald.morrison@sap.com>