docs(dev): align getting-started guides with landing page (#666)
<!-- markdownlint-disable MD041 -->
#### What this PR does / why we need it

This PR targets the _dev_ website version. It streamlines the content of
the getting started section and aligns it with the landing page (Pack →
Sign → Transport → Deploy). It contains the following changes:

- Removed the Display and Examine Component Versions section:
Introduce/Move the `ocm get` command in the _Create a Component Version_
document
- Removed the separate _Prerequisites_ page and integrate relevant
information into each corresponding guide.
- Ensured each guide is self-contained and can serve as an entry point
for the end user.

#### Which issue(s) this PR is related to
<!--
Usage: `Related to #<issue number>`, or `Related to (paste link of
issue)`.
-->
Related to
https://github.com/open-component-model/ocm-project/issues/715

---------

Signed-off-by: Bozhidara Hristova <bozhidara.hristova@sap.com>
Co-authored-by: Gerald Morrison <67469729+morri-son@users.noreply.github.com>