chore: Add note about new controllers (#671)
The current website contains Getting started guides and tutorials for
the new [OCM
controllers](https://github.com/open-component-model/open-component-model/tree/main/kubernetes/controller).
However, as they are in an early development phase, we should make the
user aware of that.

I will omit an update to the `dev` version of the website as there is
already a banner about the early development stage. This should be
sufficient.

---------

Signed-off-by: Frederic Wilhelm <frederic.wilhelm@sap.com>